Fix config.json updates blocked by other update steps failing
If the config.json is really broken, we want to be able to update
it. in this change we modify the behavior to always write down the
new version of config.json without waiting for the config update
synchronous step in the wizard.

Fixes https://github.com/beyond-all-reason/spring-launcher/issues/42